Reduce scope of dependencies
Dependencies for C and C++ on Unix (and in general) are an utter mess.
Dependencies have too much ability to change the environment you've
crafted for your software. To use dependencies, you have to include
their headers, which literally means having the compiler paste their
code into your source files. Their code may have requirements which they
need to expose, but since their code is directly embedded in yours,
those requirements are then brought to your code. If the dependency
needs to expose a non-standard function, then woe to you if you require
that function *not* to be exposed.

This is an abomination of a hack to ensure that only the relevant files
are manhandled by dependencies, reducing the scope of their destruction
as much as possible, which is to say, to only those files which actually
use the dependencies.