Build Qt with frameworks last
This is required for Github Actions, and probably for anybody, really,
since frameworks are generally old, but it's not inherently required,
and a future macos release might actually just work fine. On the other
hand I have no clue how cmake finds things on a mac, and this is really
just a magic incantation I found through a web search. For now, don't
hardcode it since it should be a user decision, but make sure Github
actions set it.